Add JavaScript code to trigger button links
In addition to recycling the Button CSS rules, I also added
a role="button" property and a draggable="false" and wrote
some quick JS to programmatically click the link when pressed
with a spacebar.

Reference HTML:
- https://design-system.service.gov.uk/components/button/#start-buttons

Using a link is semantically correct, but since we're styling like
a button, we have to replicate the expected behaviour.
In our case, that means that the link should respond to the spacebar
event (like buttons do), and the link shouldn't be draggable (because
buttons aren't).

My script tag for shimming out a spacebar press has a fairly
unscalable way of triggering the event, but it works okay for now.

Once we have to use more JavaScript, we can see if
it still works or if we need to refactor.
(eg, with a global script file.)