Remove redis session code from application
Here's the deal, @ayoajila.

We're building interface for the next couple weeks and, as far as I
can tell, we're not going to have time to get a real session hooked
into our express app.

We can add it back in later, but for now it's just code we're not using.

The point of version control is that you don't leave things you're
not using lying around forever.

So we can revisit this later.

I know full well we don't have a production-quality session management
solution at the minute, but our immediate prioritity is building an
accessible frontend for our (eventual) users.