Update focus styles for details/summary element
Updated it such that the span inside the summary has the focus style
when the <summary> element has focus.

The reason for this is that we have to display the summary element as
a list-item for the little arrow to show up in firefox, but then
the focus style spans potentially the whole length of the screen.

So the best we can do is focus the span inside.

This is pretty consistent with what happens on the GOV.UK site:
- https://design-system.service.gov.uk/components/details/