Update SIN page with user's name and SIN
Since we're going to be (eventually) creating different user
journies through the application, asking someone clicking through
to enter a correct

- access code
- SIN
- date of birth

Is a pretty big lift, even though that's how the app works.

To make it easier, I updated the hint text so that it shows the
correct value for the current user.

This is just a "dev" feature, so we will revert to static hint text
once we have a live service.

Also notable, I was originally going to add a pug filter the way that
they are documented on the website, but pug filters proper can't
take any variables. They only work with static content.

> Filters are rendered at compile time. This makes them fast, but it also means that they cannot support dynamic content or options.

This means that we needed just a pure JS function instead to display
SIN numbers in a human readable way.

Sources:
- https://pugjs.org/language/filters.html#custom-filters
- https://github.com/pugjs/pug/issues/404