Placeholder session data to get the address to work
So lots of the template code is super ugly/hardcoded, but I think
wha it points to is that we should be handling this stuff in the
controller.

So instead of 100 if checks, we should make sure that exists in the
controller and then error out if there's an issue.

Also, with the select box, we should be able to loop through this
stuff somehow.

I think for this story, we can leave this for now, but once we
loop in the user sessions, lots of this code is gone-zo.