Add validation message mixin
Included a validation message mixin in the base file as an include
statement, which means that all of our other templates have access
to it.

Prefixed all validation messages with a visually hidden "Error: "
string, also in line with the GDS guidance on validation messages.

> To help screen reader users, the error message component includes a hidden ‘Error:’ before the error message. These users will hear, for example, “Error: The date your passport was issued must be in the past”.

Also added a visually-hidden class to the repo so that we can make
content available to screenreaders but not visual users.
I have an 18F link that sums this up, and it's basically the same CSS.

Source:
- https://design-system.service.gov.uk/components/error-message/
- https://accessibility.18f.gov/hidden-content/