Move validation messages above inputs
We mulled it over with Julianna and here's where we came down.

> P: My assumption is move the error messages between the label and the input.
> J: Sometimes its more complicated but generally speaking its a good method.

The GDS guidance more or less backs this up, although it's not 1000% explicit.

> put the message in red after the question text and hint text

So it doesn't say "before the input", but that's how it is in all
of their examples.

Since our error messages show up after the page reloads, the reading
order is assumed to be

label
hint text
validation message
input

So it works for both visual users and screen reader users.

Sources:
- https://design-system.service.gov.uk/components/error-message/