Add checkPublic middleware to the app
Since we need user data to be in the app for most pages, and since
we're rapidly devving, testing, and demoing pages, we should
guarantee that we have user data in the session for all the
non-public paths.

Once this is handling real user data, we can change this middleware
to redirect everyone to the start page if they try to manually
go to a URL but haven't signed in.