Add session data to /personal/address page
Now that we have a session, we don't need to worry about
preloading the page with a fake address.

Also route to the /personal/maritalStatus page on "confirm"

Also added one more "false" check to the hasData function.
If it's an empty string, we should say there's no data there.
Added an extra test for this.