Use span instead of label for dollar sign
After a bit more research, it looks like screenreaders do funky things
when there are two labels on an input.
Also axe flags it as an accessibility violation.

It looks like technically it's not a violation, but you don't really
want to be fighting axe.

Using aria-describedby instead so that screen readers, etc will
read the name of teh question and the dollar sign.

In theory, it's a good idea.

Sources:
- https://ux.stackexchange.com/questions/13026/is-having-more-then-1-label-for-a-form-element-really-against-web-accessibility