Add schema for rrsp amount page
It turns out that express-validator is using `validator.js` under
the hood, so we have lots of methods that we can use to
do validation out of the box.

Pikced the currency validator and said that it shouldn't be negative.

Also found an interesting thing on error messages that's worth reading.

Finally, removed a dumb error key from another schema (province.province)

Sources:
- https://github.com/validatorjs/validator.js#validators
- https://design-system.service.gov.uk/components/text-input/#error-messages