Add some ✨ logic ✨ to the auth page functions
I was making a joke before to @katedee, but I think I am really
going to use middleware for this.

The problem is that we have to be able to use these pages to intercept
a few of our other pages, but only sometimes. If they have already
answered an "auth" question, then we don't want this page to
come up again.

What we can do then is check in a middleware function for a url,
and then set the redirect to res.locals and then show the auth page
instead.

Should work out.