Add doAuth middleware
One of the things we discussed with AMS was that we would use
a layered security approach where we ask people another knowledge
based question when they want to make a big change to their info,
like changing an address or marital status.

Adding some middleware that achieves this behaviour.

Expected behaviour

- if someone clicks "change address" or "change marital status", they see the /login/auth page
- they have to answer any number, not empty
- once they answer, they can change address / name
- next time they want to do it, they don't see the screen