Add stub "review your benefits" page
After entering everything, there will be one final page where users
will be able to look over what they are submitting to the CRA,
along with their estimated benefits.

Stubbed out the page for now and added one unit test.

Maybe it should be moved to another controller but this works for now.