Use different node base container in Docker file
Late Friday afternoon, our builds started magically failing to push
to Azure for some reason, and then my manual builds wouldn't run either.

After checking out the application code, it looked like there were
no proper changes to the app that would cause a previously-running
container to not run anymoer.

Also, creating containers from old commits that ran fine wouldn't
run either, so it was a bit of a head scratcher.

I figured that the thing that might have happened would have been
an update to the base container we're relying on.

So after changing from `lts-alpine` to `lts-slim`, the new
container runs again on Azure Web Apps. Thank effing goodness.

Not sure exactly the issue we were getting but it was something about
the container not starting up on time.

Here are the logs that look relevant:

2019-07-21 00:12:10.875 INFO  -  Status: Image is up to date for cdssnc/cra-claim-tax-benefits:first
2019-07-21 00:12:10.884 INFO  - Pull Image successful, Time taken: 0 Minutes and 0 Seconds
2019-07-21 00:12:10.902 INFO  - Starting container for site
2019-07-21 00:12:10.903 INFO  - docker run -d -p 19146:3005 --name claim-tax-benefits_0 -e WEBSITES_ENABLE_APP_SERVICE_STORAGE=false -e WEBSITES_PORT=3005 -e WEBSITE_SITE_NAME=claim-tax-benefits -e WEBSITE_AUTH_ENABLED=False -e WEBSITE_ROLE_INSTANCE_ID=0 -e WEBSITE_HOSTNAME=claim-tax-benefits.azurewebsites.net -e WEBSITE_INSTANCE_ID=0e632b6e7ee82a5a6d040687477d6e7796ff76bd0a2a19fdbd5dd76edfd5e90d -e HTTP_LOGGING_ENABLED=1 cdssnc/cra-claim-tax-benefits:first

2019-07-21 00:12:13.566 INFO  - Initiating warmup request to container claim-tax-benefits_0 for site claim-tax-benefits
2019-07-21 00:16:24.851 ERROR - Container claim-tax-benefits_0 for site claim-tax-benefits did not start within expected time limit. Elapsed time = 251.284957 sec
2019-07-21 00:16:24.862 ERROR - Container claim-tax-benefits_0 didn't respond to HTTP pings on port: 3005, failing site start. See container logs for debugging.