Add /commencer route so that we can load the site in french
We haven't really solved the problem that all our URLs are in
English, but that doesn't mean that we have to make people
type a query parameter just to see the site in french from the jump.

Added a new route that redirects to the start page in French.

Pretty simple, works good enough for now.