Add tests
One big change here is that given how the default placeholder of the amount text box is zero we accept null and empty as valid values for either provincial or federal amounts. This is so that if a user enters something in only one box we don't give an error message.
Currently if both boxes are empty or null we also accept these as both being 0.
This is different than other pages and we may want to look into changing this inconsistency in the future.