Conditionally add details polyfill for older browsers
So there's a few decisions in here.

- Adding a polyfill

<details> elements don't work in IE11, but they're a nice semantic
solution for what we want to do: conditionally display information.

The easy way to fix this without a lot of finagling is to include a
polyfill. I've picked the `details-element-polyfill` on npm because
it's the most popular one.

- Using a CDN

We could include this as a local package and then include it
from node_modules, but I didn't. CDNs are pretty popular and
tend to work. Unless this is gonna get blocked by CRA @ayo,
I think this works.

- Using a conditional statement with document.write()

There are more ways to do this that are discussed at length in this
this other article: https://philipwalton.com/articles/loading-polyfills-only-when-needed/

However, this was cheap, easy, and it works, so I didn't want to overdo it.

Source:
- https://stackoverflow.com/questions/45509701/elegant-way-to-include-es6-promise-for-ie