Only include polyfill on pages that need it
Our conditional include was asking for the polyfill to be loaded on
every pageload in older browsers.

However, most of our pages don't have a summary element, so we don't
need that extra request most of the time.

So this extra call looks for a `details` element in the HTML before
loading the polyfill.