Insource the details polyfill
Thinking about it, using a CDN might cause weird problems because
@ayoajila has been telling me there are weird proxy issues going on.

It also means that our container isn't self-contained anymore if it
relies on an external request.

So, I just copied it into a /js directory. I doubt it gets updated
very often and if we need to update it we can just copy in the latest
version.

I figured we don't really need to include this in our package.json
as a dependency.