Blank is valid answer to currency questions
During usability testing some users had trouble with pressing "continue" because the placeholder was showing 0. Having a default value of 0 also made the users type in numbers incorrectly.
This patch changes it so that all fields that follow currencySchema can accept blank (and null).
Since we no longer need to have the logic separately for political contributions we just move them to use the currency schema.
For the /auth/ page the test have been changed to accept an empty value since it could be empty.