Add "secure" cookie setting to the app if we think we're on prod
If we're running with "npm start", and we have a GITHUB_SHA
environment variable, we're very likely to be in production, which
means we're running on Azure.

If that's the case, we're on https and the cookies should have the
"secure" flag, which means they won't be served on http.

If misconfigured, this could mean that we're not saving data locally
while developing, but it seems unlikely to me, and if the devs on the
project know about it, then it should be fine.