Add 'role="button"' to all of our link buttons
When you shim a link as a button, you want to put a 'role="button"'
attribute onto it otherwise screenreaders won't know that you
want it to be perceived as a button.

Didn't notice earlier that we were missing the 'role="button"'
attribute on a bunch of our buttons earlier, so I added them in.

Also changed our CSS and JS selectors to look for the role attribute
because that's what's semantically important.