Remove the "secure" flag on cookies
Unfortunately, setting secure cookies for us in production doesn't
actually work because I am pretty sure our app is sitting behind
a proxy of some kind or other.

I noticed that when ‘secure’ cookies were set and the app was running locally, you don’t see any “response cookies” in Chrome's "network" tab.
Same thing with the "prod" version of the app.

When we upload our app to app service, we click some option that says “only serve traffic over HTTPS”, which means probably our app is running on HTTP behind a proxy.

This means the secure option is set, but the cookies don't get
set at all. So the solution for now is to turn them off.