Add validation for all fields
Tacked on the "birthdate matches" validation onto the day field.

Since it's going to trigger way more often than the others, we
need to put some logic into the template (or the route) to ignore
this error if we have any other errors existing.

The error message validation stuff is so obviously broken that it
will have to be bespoke for this page, so we're pretty okay to
manage this page as a special case without interfering with our
pattern we use on other pages.