Add fieldset and legend to date of birth fields
A fieldset is used to group conceptually related form fields,
which is exactly what we are doing with the date of birth stuff.

This is also recommended by the GOV.UK Design System:

> The 3 date fields are grouped together in a <fieldset> with a <legend> that describes them, as shown in the examples on this page. This is usually a question, like ‘What is your date of birth?’.

As a side note, I noticed we were using a .fieldset class to style
our fieldsets, and I removed the classname in favour of just styling
the elements themselves.

Source:
- https://design-system.service.gov.uk/components/date-input/