Refactor hasData and added tests
It turns out that our hasData function wouldn't actually work
how we were expecting it to. Specifically, the `.every()` function
will _always_ return true or false, which means that if we want
to return the actual value, we have to wait for the result of
the .every() loop to finish and then look through the string for the
original object.

We were also mutating the object by reassigning it, which is kind of
another thing, so I used Object.assign to create a copy of the
original passed-in object for us to mutuate.

Let's see what LGTM says about it.

Source:
- https://stackoverflow.com/questions/6491463/accessing-nested-javascript-objects-with-string-key
- https://stackoverflow.com/a/43849204
- https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/every