Create 2 actions files: 1 to test and 2nd to test and deploy
Maybe it is my fault and I can't figure it out, but it looks like,
based on the documentation, that the new github actions don't support
filtering based on branch names in an easy way.

ie, What I want to do is run the tests on all branches, and then
run the build + deploy on ONLY the master branch.

I can't figure out the way to do that with GitHub's new syntax, so
I've come up with this new 2-file system that seems obviously bad
but whatever.

There is an "if" syntax, but I think I would have to write
"if: branch is master" like 8 times, so that seems bad.

So here's what I've done.

File 1 will run the tests on all branches EXCEPT master.
File 2 will run tests + deploy only on master branch.

If there is a better way to do this, I am all ears.

Source:
- https://help.github.com/en/articles/workflow-syntax-for-github-actions
- https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstepsif