Use unicode charater sequence for nonbreaking space in French content
Up to this point, we've been using non-breaking spaces that
are visually identical to normal spaces. But we should have some
kind of space that is visually distinct.

Adding the unicode character sequence is a better solution.

Note: there is a mention in the discussion that we should be using
half-spaces but we haven't been doing that so far.

Source:
- https://github.com/kazupon/vue-i18n/issues/318
- http://www.fileformat.info/info/unicode/char/00a0/index.htm