Whitelist redirect URLs
LGTM pointed out that our redirects are pretty loosey-goosey,
basically, we'll redirect you to any url that's in the form request.

Since we now have an array with all the url paths, we can make sure
that the redirect url is one we actually know about before
sending people there. Pretty small change, but @johnO will be happy.