Use morgan logger with JSON format
Removed the winston logger entirely in favour of just using morgan,
which is pretty easy to configure and work with.

Also sound a snippet online showing how to do JSON logs, which are
easier to search through if we're going to have to be doing that
in the near future (which I think we will do).

Created a smaller format for dev and a larger format for production,
which we can see by using 'npm start'.

Winston lets us create lots of different log files and stuff, but
since we're doing a container deployment, we just want everything
going to stdout, so we don't have that complex a use case.

Note: not saving any user IP info because that's a can of worms we
don't want to open anytime soon.

Sources:
- https://github.com/expressjs/morgan
- https://gist.github.com/cgmartin/913bb12097ff07132597