Only run "deploy" job if on master branch
Previously, I had thought there were only two ways to check for
specific branches:
- running or ignoring an entire workflow file
- running a specific step of a job

So that meant that I created two workflow files, one for the master
branch, and one for all other branches.
In the master branch workflow, I ran test + deploy actions, and in
the workflow for non-master branches, I copied all the "test" code.

However, I just learned that you can check for specific branches
in the root of a job itself, which is what I was looking for originally.

So now we can delete the file that was only for non-master branches
and add a conditional to the "deploy" job checking for the master
branch. Saves us a bunch of copy and paste.

Sources:
- https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idif
- https://help.github.com/en/articles/workflow-syntax-for-github-actions#jobsjob_idstepsif
- https://help.github.com/en/articles/contexts-and-expression-syntax-for-github-actions#functions
- https://help.github.com/en/articles/virtual-environments-for-github-actions#default-environment-variables