Replace cookie-session with memorysession
For some reason, our cookie session has not been saving values
properly (either appearing to keep them or drop them) and this has
caused us much angst.

After a bit of investigation, I figured out that that cookie
session was actually getting dropped on _every new pageload_,
which was alarming and seemed to imply we were doing something wrong.

Not so, however. It wasn't our middleware, and it wasn't that we
implemented something wrong, it was actually that cookies have a maxium
size, and we hit it some time ago.

Our session was getting dropped consistently on every new page, but if I removed
a section from the api/user.json file, it worked fine.
For example, removing the 'benefits' section at the bottom would fix everything.

What's happened is that over time, our session data has gotten bigger and bigger,
and then once the cookie stopped being saved, the middleware function that
would add the new session in would just take over and load it in per request.

That meant that we always _appeared_ to have a session, but it was actually
getting dropped every time. So we weren't able to save things or w/e. It
was hard to debug.

Picked 'memorystore' as they claim to have fixed the memory leak issue with
the default session memory storage:

> express-session full featured MemoryStore module without leaks!

Also:

Ran into a problem assigning straight to the session variable

```
req.session = user //causes an error
```

So the fix was to assign to a key

```
res.session.personal = user.personal
```

Which means we're now looping all keys in user to assign the session.
Not a big deal, just noting it for future.

Sources:
- https://www.quora.com/What-Is-The-Maximum-Size-Of-Cookie-In-A-Web-Browser
- https://github.com/roccomuso/memorystore
- https://davidburgos.blog/expressjs-session-error-req-session-touch-not-function/