Save responseId to the session and log it out per request
On the upcoming survey, people are going to load the URL inside of
an iframe with the ResponseID query parameter.

If we see the query parameter, we want to save it to the session and
then log it out every time, so that we can see someone's request paths
as they move through the application.