Preserve query parameters from / to /start page
Since our start page is the main entry point to the site, we might
might be in situations where we need the query params from the root
url to the /start url, so passing the query params through seems like
a fine idea.

Wrote a new unit test to confirm behaviour.