Remove max header size
Removed the max header size from the `npm start` command.

When we were using cookie session earlier, we started to get weird
400-level errors because our headers were getting huge because
we were storing so much information in the cookies.

Bumping the header size resolved this for us, but it's caused problems
running the app cross-platform (for example, on github actions or
on windows).

Now that we've moved away from cookie session, we don't need this
instruction anymore because we don't have very large headers, so
let's get rid of it.

Also added "node" command to eslint so that it's more resilient.

Sources:
- Adding the option: https://github.com/cds-snc/cra-claim-tax-benefits/pull/92
- Removing cookie-session: https://github.com/cds-snc/cra-claim-tax-benefits/pull/292