Rewrite SIN schema to not change the user's submitted value
Previously, we would sanitize the input SIN to remove spaces or
hyphens before validating it.

Unfortunately, sanitizing it would also mutate the submitted
value, so repopulating the form would show them a validation error
and then a value they probably never typed in.

Like if I submitted "123 456 78", it would come back as "12345678".

So it's telling me I'm wrong, but it also rewrote what I sent in.
That's not really fair.

Changed the SIN validation to _only_ use our custom validation method
and send back the same error messages. The code is uglier but
it has tests now, so should be okay.

Behaviour should be the same as previously.

Also, added an underscore to my internal functions because they're
supposed to be private.

Inspired by @OmgImAlexis' comment 👇

Source:
- https://github.com/express-validator/express-validator/issues/552#issuecomment-489400183