Refactor the doYesNo middleware
Previously we always had "yes" or "no" pages in front of "amount" pages,
so I wrote a pretty basic middleware that assumed this would always
be the case.

Now that it is no longer the case, we have weird functions that do
a bunch of copy+paste stuff and we gotta get rid of that nonsense.

Updated the doYesNo function so that it can handle nulling out
more than just the "amount" variable, and it's no longer hardcoded
to use the path + "/amount" url that we were assuming before.

Easy.