Move code around so that cheerio is a test dependency again
Since moving `extractCSRFToken` function to the `/utils/index.js` file,
we were technically relying on the 'cheerio' package for the prod
version of the application — not because the `extractCSRFToken` is
used in the prod version of the app, but because other functions
in `/utils/index/js` _are_ getting called, and when trying to call them,
the app would crash because `/utils/index.js` file was missing an import.
(Cheerio was not being installed when running `npm install --production`.)

The quick fix was to promote the cheerio dependency into our prod
dependencies, but it's 100% _not_ a prod dependency so this PR fixes that.

This PR creates a new file from which we can import our extractCSRFToken
test function, and has a few basic tests for it in the same file.

Having done that, we can put the cheerio lib back in as a 'devDependency',
which is where it belongs.

Easy.