Add option for currencySchema to fail empty values
Previously, currencySchema would allow an empty submission (it would
return a "0" when it got an empty string.

However, that's not the behaviour we want for the security questions,
so added an option to fail it on empty POST requests.

Also rewrote the taxReturn tests so that I can duplicate them
more easily.