Add explicit max-height to FIP svg
On IE on Windows 7 and under, the SVG in the header has a height
that is far too large, even though it visually shows up normally.

Checked how Canada.ca solves this problem and they set a max-height
at 40px, which seems to work and isn't too complicated.

Looked up some stackoverflow questions but they mostly had complicated
fixes so gonna go with this instead.