Add client-side JS to remove the trailing URL hashes before a form submit
Anchor links that are part of the URL are preserved when we submit a form, whether it succeeds or fails.

Examples:
- Hitting the "skip to main content" link on the 'code' page will add `#content` to the end of the URL, and then submitting either a right or wrong code won't get rid of it
- Clicking a validation error link in the error summary will put an anchor link in the URL (eg, `#code`), and then it stays there on the next page as well.

With real links we don't see this behaviour: ie, from `/start#content`, we can click "Start now" and the `#content` part of the URL disappears.

Even allowing that this seems to be the default behaviour (we never modified it), I would say that this is unexpected an unneeded more often than it's helpful.

What we should do instead is always strip the `#` anchor part of our URLs whenever we redirect on a POST route, whether it fails or passes.

It looks like this is expected behaviour for a different use case:
- https://stackoverflow.com/questions/2286402/url-fragment-and-302-redirects

Basically, if a page is moved but you had a table of contents, the hash links should
still work. This isn't the context we have.