Use node:lts-alpine as base Docker image
Early on, we were using `lts-alpine` as our base image, but then
one weekend the builds started failing mysteriously.

At that point, I swapped out the base container to `lts-slim`,
which solved our problem.

Long-term, the `lts-alpine` containers are better for us to be using,
so changing it back now that the problem appears to be resolved.

Source:
- 5ac889103b20c3ff77d572f0cff0566f4873c9b3