More resilient method for removing hashes from URLs
Our previous code for removing hashes from URLs wasn't working
in IE11, but this one does a check for non-compliant browsers,
as well as doesn't remove the last part of the URL on IE11
(that's why we were getting the 404).

This fixes a bug where trying to submit a post request on IE11
would cause a 404 error.

Source:
- https://stackoverflow.com/questions/1397329/how-to-remove-the-hash-from-window-location-url-with-javascript-without-page-r/5298684#5298684