Return row values camelCased so that they match everything else
The idea here is to keep all the database row stuff inside of the
db.js file, not have it in our controller files.

We call it firstName everywhere, so keep that consistent in all of
the application code.

Also update the db tests to reflect the changed format of
the row names.