Update address logic in the view
Since adding 'en' and 'fr' keys, our address template is displaying
the address incorrectly when there's no apartment number.

The 'if' condition is not aware of the changes so it will always
return true, inserting a hyphen before the street address every time
eg, "-180 Lisgar"

Furthermore, the "else" case doesn't know about the language key
either, so it prints out the object.
eg, "[Object object]"

This change to the template fixes both issues.