Use `is65` function to check if someone _was_ over 65
When we are asking people's ages, we want to know if they were
taht age for a given tax year, not for the current year.

If someone turned 65 in January, it doesn't affect their taxes
from the previous year. We want to know if they were 65 on December 31
of the previous year.

This means we can remove the 'getDateDelta' function. Knowing
someone's actual date on the day they're using the service
isn't necessary, so removed the 'getDateDelta' function entirely.