Check the dateOfBirth on the /confirm-income page
When someone's going through the eligibility flow, we ask them to
tell us if they are over 65, but we don't really know if they are.

Later, as one of the last steps in the service, we ask people to
confirm their income again, and we show different income amounts
depending on if they are above or below 65.

We were using their same answer they gave to the eligibility question
to check if they are over 65 or not, but that was actually the wrong
way to go about it.

Similar to the "senior transit page", we know what age someone is
because we have their date of birth back from the API.

This pull request changes the logic to show the income amount based
on their date of birth, and not based on their previous
answer they gave during the eligibility flow.