Add hidden text to "learn more" links on benefits table
Having a bunch of links on the page that all say the same thing
is really terrible for screen readers, so we want to put in some
contextual information for which benefit each of these links refers to.

Created a separate key for the link text because the french articles
change between "le" and "la", such that a generic intro (eg, "sur le")
wasn't going to cut it.