Add update process for `details-element-polyfill.js`
The details element polyfill we have is different than all of our
other dependencies.
It doesn't get included in most browsers, and even then it doesn't
get used on most pages, so we don't want it included 100% of the time.

This file doesn't get updated very often so we're okay to keep doing this.
Just adding a note so that we don't lose it.