Add page of code review guidelines
I went over a bunch of different sources for this.
I tried to split the differnece and come up with something that is
useful for reviewers to read, both in terms of the tone of reviews and
in the more technical terms of what to look for.

Here are my sources:

- https://github.com/alphagov/styleguides/blob/master/pull-requests.md
- https://github.com/cds-snc/.github/blob/master/CODE_OF_CONDUCT.md
- https://github.com/alphagov/styleguides/blob/master/git.md
- https://github.com/thoughtbot/guides/tree/master/code-review

- https://www.evoketechnologies.com/blog/code-review-checklist-perform-effective-code-reviews/
- https://www.codementor.io/blog/code-review-checklist-76q7ovkaqj
- https://www.michaelagreiler.com/code-review-checklist/