Add Architecture decision record about our choice of Express
Since the beginning, there has been discussion of whether or not
to use Angular, and I am pretty sure folks from CRA will want to know
why we went with Express instead.

I did a spike about it and then a writeup as to why we did what we
did, so let's just add it to the repo.